requirements.txt
numpy>=1.21.0
torch>=2.0.0
matplotlib>=3.5.0
pillow>=9.0.0
imageio>=2.25.0
imageio-ffmpeg>=0.4.8
tqdm>=4.64.0
